# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/rooms": {
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      },
      "get": {
        "operationId": "RoomsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
    "/api/rooms/available": {
      "get": {
        "operationId": "RoomsController_findAvailableRooms",
        "summary": "Find available rooms",
        "description": "Returns a list of available rooms based on search criteria",
        "parameters": [
          {
            "name": "checkInDate",
            "required": true,
            "in": "query",
            "description": "Check-in date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-09-25",
              "type": "string"
            }
          },
          {
            "name": "checkOutDate",
            "required": true,
            "in": "query",
            "description": "Check-out date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-09-30",
              "type": "string"
            }
          },
          {
            "name": "guests",
            "required": true,
            "in": "query",
            "description": "Number of guests",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "roomType",
            "required": false,
            "in": "query",
            "description": "Room type",
            "schema": {
              "example": "standard",
              "type": "string"
            }
          },
          {
            "name": "roomTypeId",
            "required": true,
            "in": "query",
            "description": "Type of room",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available rooms"
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stellar Backend API",
    "description": "Stellar Backend API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "number",
            "description": "Identificador único del tipo de habitación, se genera automáticamente",
            "example": "1"
          },
          "beds": {
            "type": "number",
            "description": "Número de camas en la habitación",
            "example": 2
          },
          "maxOccupancy": {
            "type": "number",
            "description": "Ocupación máxima de la habitación",
            "example": 4
          },
          "oceanView": {
            "type": "boolean",
            "description": "Si la habitación tiene vista exterior/oceánica",
            "example": true
          },
          "basePrice": {
            "type": "number",
            "description": "Precio base por noche en dólares",
            "example": 60
          }
        },
        "required": [
          "typeId",
          "beds",
          "maxOccupancy",
          "oceanView",
          "basePrice"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}